Removed most analyzer warnings.

+++ src/common/maptypes.h
@@ -674,7 +674,8 @@ struct MapUserAircraft
   }
 
   explicit MapUserAircraft(const atools::fs::sc::SimConnectUserAircraft& aircraftParam)
-    : MapBase(map::AIRCRAFT, aircraftParam.getObjectId(), aircraftParam.getPosition()), aircraft(aircraftParam)
+    : MapBase(map::AIRCRAFT, static_cast<int>(aircraftParam.getObjectId()), aircraftParam.getPosition()),
+    aircraft(aircraftParam)
   {
   }
 
@@ -704,7 +705,8 @@ struct MapAiAircraft
   }
 
   explicit MapAiAircraft(const atools::fs::sc::SimConnectAircraft& aircraftParam)
-    : MapBase(map::AIRCRAFT_AI, aircraftParam.getObjectId(), aircraftParam.getPosition()), aircraft(aircraftParam)
+    : MapBase(map::AIRCRAFT_AI, static_cast<int>(aircraftParam.getObjectId()),
+              aircraftParam.getPosition()), aircraft(aircraftParam)
   {
   }
 
@@ -726,7 +728,8 @@ struct MapOnlineAircraft
   }
 
   explicit MapOnlineAircraft(const atools::fs::sc::SimConnectAircraft& aircraftParam)
-    : MapBase(map::AIRCRAFT_ONLINE, aircraftParam.getObjectId(), aircraftParam.getPosition()), aircraft(aircraftParam)
+    : MapBase(map::AIRCRAFT_ONLINE, static_cast<int>(aircraftParam.getObjectId()),
+              aircraftParam.getPosition()), aircraft(aircraftParam)
   {
   }
